refactor of build to include every time the GratefulDeadConcerts in the dist
